Finally make popups work and disable links in preview mode.
The key problem with iframes is that the dimensions have to be set
explicitly, auto geometry does not work. Next problem was that a
textarea element does not fire a resize event. We have to emulate one.
Next: in preview mode wiki-links have to be disabled, otherwise the
browser just follows the link and all the changes are lost.